requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
